Namespace client errors
Redis uses the convention of having the first letters of the response
string up to the first space, denote the category of the error:
https://redis.io/topics/protocol (see "RESP Errors" section). We use the
same convention to communicate the type of errors to the client, so that
they can treat them accordingly. The reason for this is that a consuming error
might need different handling than a producing error, depending on the
application business logic.

We might introduce more categories in the future, but for now it's just
producing and consuming errors, denoted by the "PROD" and "CONS"
prefixes respectively.