Make Client responsible for closing its conn
Prior to this, closing client connections was done in a separate
step outside of `client.Close()`.

However it makes sense to move this responsibility to Client, the
higher-level abstraction that already contains conn in its internal
state.

This change does not alter the actual closing sequence:
consumers/producers are closed first and then the client connection.

Closes #36.