Shutdown process overhaul
This patch aims to minimize the downtime for producers during server
upgrades and generally make the shutdown process more robust.

The flow is now the following:

1) shutdown signal is received
2) stop accepting new consumers and close clients with at least 1
consumer
3) stop accepting new producers and close rest of the clients
4) shutdown and restart

This way, producer downtime is reduced to the time elapsed between (3)
and (4), which should be less than a second. Nevertheless, clients
should still have sane retry defaults configured anyway, because some
downtime will still occur.

This patch also changes the client IDs (Client.ID) to be unique per-client.
These are internal and should not affect the consumer logic, since we
introduce Client.consID and we use that for consumer IDs.