Properly return parse errors to clients
Previously, errors returned from go-redisproto parser that were not of
type `redisproto.ProtocolError`, were not communicated to clients. Clients
would just see that rafka closed the connection but without a reason.

For example, this happened when we got a `redisproto.InvalidBulkSize` error
due to very large messages sent to rafka.

From now on we return the respective error strings to clients, despite the
type of the error.

Before, rafka-rb returns:

    > producer.produce("foo", "a"*65*1024)
    Redis::ConnectionError: Connection lost
    (ECONNRESET)

after this change it returns:

    > producer.produce("foo", "a"*65*1024)
    Rafka::CommandError: Invalid bulk size