Add support for the MONITOR command
MONITOR is a debugging command that streams back every command processed
by the Redis server[0]. Adding the ability to see all the requests
processed by the Rafka server will be quite useful in order to spot bugs
easier in a Rafka application.

To enable this functionality, we add an extra channel to the `Client`
struct, the `monitorChan`. A separate goroutine that is spawned at the
client's creation time, will stream any data written to that channel on
the client's socket.  On server-side, we maintain a map of the clients'
monitoring channels, named `monitorChans`. Each time a MONITOR command
is triggered, the respective client's channel is appended to that map.
Another goroutine named `monitorHandler` is in-charge of writing every
command received by the Rafka server to the respective clients' channels
via the main `srvMonitorChan` channel.

[0]: https://redis.io/commands/monitor