Drop support for multiple Consumers per Client
Currently, a `Client` instance is allowed to have more than one
`Consumer` instances associated with it. As the respective ddoc details
(`design-rafka-rethinking.rst`), there is no actual reason to maintain
this one-to-many relationship between Clients and Consumers which among
others corresponds to a lot of extra logic in the source code.

This commit drops support for multiple `Consumer` instances per
`Client`; from now on, a `Client` instance can only be mapped to a
single `Consumer`. In order to drop this functionality, the following
modifications have been applied:

- drop the `Client.consumers` map completely
- drop the `Client.consByTopic` map completely
- (temporarily) introduce the `consID` field on Client struct
- replace the `ConsumerByTopic` function with the `fetchConsumer`
  equivalent

As mentioned, the `consID` field is temporary; it is mainly used to
facilitate the transition to a `ConsumerManager`-free Rafka server.