Drop the consumer_manager module completely
This commit introduces some major changes in Rafka internals, as the
`design-rafka-rethinking.rst` ddoc details. In short, the
`consumer_manager.go` module and all the respective logic, is completely
dropped from the source code. Subsequently, the handling of the
`Consumer` instances will be shared between the `Client` and `Server`
instances.

To do so, we've had to make the `Client` struct aware of the server's
`Context` as well as the Rafka configuration file. Since `Client`
becomes the authoritative handler of `Consumer` objects, those two extra
fields are essential to properly register and de-register a `Consumer`
instance via the `Client` object.

Moreover, the temporary `consID` Client's field is removed and replaced
by the `consumer` one, which is a reference to the `Client`'s `Consumer`
object. Also, a new function is introduced in `Client`, named
`registerConsumer` which will be used to create a new `Consumer`
instance upon request. The `Consumer` struct is extended with a new
field containing the parent's cancel context (aka `Client`), which will
allow us signal a `Consumer` instance about the Client's termination; a
task which was performed until now by the `ConsumerManager` handler.