Wait for all consumers to Close upon shutdown
When the server is requested to shutdown we need to wait for all active
consumers to close prior to stopping the server, a task which was among
ConsumerManager's responsibilities. Now that the respective module is
dropped, we make the server the authoritative controller of this task
via a `sync.WaitGroup` struct.