server: Stop accepting new Consumers during shutdown
A server shutdown operation may last longer than expected in case the
Rafka server is loaded. During this interval we don't want to register
any new Consumers to Rafka. This commit introduces a new flag in server
struct named `teardown`, which denotes whether a shutdown request is
in-progress. If so, no new consumers will be registered and an
appropriate message will be returned to the client. Since the `teardown`
flag is shared among the Server and its Client's, we had to guard its
references using a `sync.RWMUtex`.

Note:

  In contrast to Consumer's handling, we allow new Producers to be
  registered while the server shuts down. Producers are way more
  lightweight comparing to Consumers, and so we can start and stop them
  freely. This way we, we also descrease the downtime interval of
  registered Producers in Rafka.