producer: Do not mark ignorable events as errors
Until now we are treating producer connection errors as message delivery
errors which is raising false alarms.

As per the example [1] provided by the developers, we simply log these events
from now on, but also we keep track of unrecognised events separately.

This commit adds a separate stats counter for the event errors which
can either be ignored or be used by the applications. In any case, we do
not just log them, but also allow for alerting based on this.

[1] https://github.com/confluentinc/confluent-kafka-go/blob/master/examples/producer_channel_example/producer_channel_example.go#L48-L66